# openapi.yaml
# OpenAPI Specification
openapi: "3.0.0"
info:
  version: 0.1.0
  title: Project Manager - Backend
  description: Backend API for the Projekt Manager
  contact:
    name: Cloud Administrator
    email: cloud@wwu.de
  license:
    name: MIT
    url: https://choosealicense.com/licenses/mit/
paths:
  /healthz:
    get:
      description: Health check endpoint
      operationId: getHealth
      responses:
        '200':
          description: Everything OK
          application/json:
            schema:
              type: string
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/project/{name}:
    get:
      description: Get Project by Name
      operationId: GetProject
      parameters:
        - $ref: '#/components/parameters/name'
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/project'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      description: Create or Update a Project
      operationId: EnsureProject
      parameters:
        - $ref: '#/components/parameters/name'
      requestBody:
        description: New or updated Project
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/project'
      responses:
        '201':
          description: Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/project'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      description: Delete a Project
      operationId: DeleteProject
      parameters:
        - $ref: '#/components/parameters/name'
      responses:
        '200':
          description: Project deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'

components:
  parameters:
    name:
      name: name
      in: path
      description: Name of the Project
      required: true
      schema:
        type: string
        minLength: 1
        example: example_name
  schemas:
    project:
      type: object
      required:
        - name
        - config
        - status
      properties:
        name:
          type: string
          readOnly: true
          example: example_name
        description:
          type: string
        faculty:
          type: integer
          example: 0
          description: Number of the faculty (https://www.uni-muenster.de/wwu/fak_fb/)

    error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          example: "Not Found"